+++ quantum/matrix.h
@@ -73,6 +73,7 @@ void matrix_init_user(void);
 void matrix_scan_user(void);
 
 #ifdef SPLIT_KEYBOARD
+bool matrix_post_scan(void);
 void matrix_slave_scan_kb(void);
 void matrix_slave_scan_user(void);
 #endif
